requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-openai==0.0.2
faiss-cpu==1.7.4
numpy==1.26.1
pandas==2.1.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
torch==2.0.1
python-multipart==0.0.6
jinja2==3.1.2